fix: Remove component method definition out of useAtlasmap.tsx and into its own component support utility. Fixes: #1520
By mirroring the existing UI components directory it will be easier to make sure we don't miss
anything.  It additionally reduces the complexity of useAtlasmap.tsx.
